Add an explicit cast to shut up warnings from g++ 3.4, addressing bug #13269

+++ src/unit_id.cpp
@@ -22,7 +22,7 @@
 namespace n_unit {
 	id_manager id_manager::manager_;
 
-	id_manager::id_manager() : next_id_(0), fake_id_(-1)
+	id_manager::id_manager() : next_id_(0), fake_id_(static_cast<size_t>(-1))
 	{}
 
 	id_manager& id_manager::instance()
@@ -58,7 +58,7 @@ namespace n_unit {
 
 	void id_manager::reset_fake()
 	{
-		fake_id_ = -1;
+		fake_id_ = static_cast<size_t>(-1);
 	}
 
 	void id_manager::clear()
